% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/demagrittr.R
\name{demagrittr}
\alias{demagrittr}
\title{decompile magrittr's syntax to normal R's syntax}
\usage{
demagrittr(expr_, eval_ = FALSE)
}
\arguments{
\item{expr_}{expression with magrittr functions such as "\%>\%"}

\item{eval_}{evaluate the result if value is TRUE}
}
\description{
decompile magrittr's syntax to normal R's syntax
}
\examples{
demagrittr(x \%>\% f)
demagrittr(x \%>\% f(y))
demagrittr(x \%>\% f \%>\% g \%>\% h)

demagrittr(x \%>\% f(y, .))
demagrittr(x \%>\% f(y, z = .))

demagrittr(x \%>\% f(y = nrow(.), z = ncol(.)))
demagrittr(x \%>\% {f(y = nrow(.), z = ncol(.))})
}

