% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demagrittr_source.r
\name{demagrittr_source}
\alias{demagrittr_source}
\title{Convert R's source file to non \%>\% code.}
\usage{
demagrittr_source(in_, out_)
}
\arguments{
\item{in_}{file path of input.}

\item{out_}{file path of output.}
}
\value{
no return. side effect on a file.
}
\description{
read file and convert the code which uses %>%
  to flatten code
}
\examples{
tmp_dir <- tempdir()
in_path <- paste0(tmp_dir, "test_in.r")
out_path <- paste0(tmp_dir, "test_out.r")

writeLines(
"
x <- data.frame(a = 1:5, b = 6:10)
y <- x \%>\%
  select(b) \%>\%
  filter(b >= 8)
", in_path)

demagrittr_source(in_path, out_path)

system(sprintf("cat \%s", in_path))
system(sprintf("cat \%s", out_path))

}

