% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demagrittr.R
\name{demagrittr}
\alias{demagrittr}
\title{Decompile magrittr's syntax to normal R's syntax}
\usage{
demagrittr(expr_, eval_ = FALSE)
}
\arguments{
\item{expr_}{expression with magrittr functions such as "\%>\%"}

\item{eval_}{evaluate the result if value is TRUE}
}
\description{
Semantics of \code{\%>\%} is eager evaluation, the example of magrittr
 is not accurate. \code{x \%>\% f} is not equivalent to \code{f(x)} but equals to
 \code{{tmp <- x; f(tmp)}} or \code{{force(x); f(x)}}. demagrittr() converts magrittr's syntax to such eager
 evaluation.
}
\examples{
demagrittr(x \%>\% f)
demagrittr(x \%>\% f(y))
demagrittr(x \%>\% f \%>\% g \%>\% h)

demagrittr(x \%>\% f(y, .))
demagrittr(x \%>\% f(y, z = .))

demagrittr(x \%>\% f(y = nrow(.), z = ncol(.)))
demagrittr(x \%>\% {f(y = nrow(.), z = ncol(.))})

}

