% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demagrittr.R
\name{demagrittr}
\alias{demagrittr}
\title{Decompile magrittr's syntax to normal R's syntax}
\usage{
demagrittr(expr, is_NSE = TRUE, mode = c("eager", "lazy", "promise"))
}
\arguments{
\item{expr}{expression with magrittr functions such as "\%>\%"}

\item{is_NSE}{if TRUE, expr is not evaluated.}

\item{mode}{choose one of "eager", "lazy", and "promise".
Pipe streams are formed along with such mode.}
}
\description{
Semantics of \code{\%>\%} is eager evaluation, the example of magrittr
 is not accurate. \code{x \%>\% f} is not equivalent to \code{f(x)} but equals to
 \code{{tmp <- x; f(tmp)}} or \code{{force(x); f(x)}}. demagrittr() converts magrittr's syntax to such eager
 evaluation.
}
\examples{
demagrittr(x \%>\% f)
demagrittr(x \%>\% f(y))
demagrittr(x \%>\% f \%>\% g \%>\% h)

demagrittr(x \%>\% f(y, .))
demagrittr(x \%>\% f(y, z = .))

demagrittr(x \%>\% f(y = nrow(.), z = ncol(.)))
demagrittr(x \%>\% {f(y = nrow(.), z = ncol(.))})

demagrittr(x \%>\% f \%>\% g, mode = "eager") # default
demagrittr(x \%>\% f \%>\% g, mode = "lazy")
demagrittr(x \%>\% f \%>\% g, mode = "promise")

}
