% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demagrittr.R
\name{demagrittr}
\alias{demagrittr}
\alias{demagrittr}
\title{Decompile magrittr's syntax to normal R's syntax}
\usage{
demagrittr(expr, is_NSE = TRUE, as_lazy = FALSE)
}
\arguments{
\item{expr}{expression with magrittr functions such as "\%>\%"}

\item{is_NSE}{if TRUE, expr is not evaluated.}
}
\description{
Semantics of \code{\%>\%} is eager evaluation, the example of
 magrittr is not accurate. \code{x \%>\% f} is not equivalent to
 \code{f(x)} but equals to \code{{tmp <- x; f(tmp)}} or
 \code{{force(x); f(x)}}.
 demagrittr() converts magrittr's syntax to such eager evaluation.
}
\examples{
demagrittr(x \%>\% f)
demagrittr(x \%>\% f(y))
demagrittr(x \%>\% f \%>\% g \%>\% h)

demagrittr(x \%>\% f(y, .))
demagrittr(x \%>\% f(y, z = .))

demagrittr(x \%>\% f(y = nrow(.), z = ncol(.)))
demagrittr(x \%>\% {f(y = nrow(.), z = ncol(.))})

}

